% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{attenborough_palette}
\alias{attenborough_palette}
\title{Get an Attenborough color palette}
\usage{
attenborough_palette(name, n = NULL)
}
\arguments{
\item{name}{The name of the palette}

\item{n}{Number of colors to return (optional)}
}
\description{
Get an Attenborough color palette
}
