% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{attenborough_palettes}
\alias{attenborough_palettes}
\title{Attenborough Color Palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
attenborough_palettes
}
\description{
A collection of color palettes inspired by nature documentaries, narrated by Sir David Attenborough.
}
\keyword{datasets}
