% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{display_attenborough_palette}
\alias{display_attenborough_palette}
\title{Display an Attenborough color palette}
\usage{
display_attenborough_palette(name)
}
\arguments{
\item{name}{The name of the palette}
}
\description{
Display an Attenborough color palette
}
